added some rpi3 stuff and fixed the port system

+++ include/generic/io_nucleon.hpp
@@ -21,20 +21,20 @@ public:
   void reset(void) const;
 
   // Get a uint8_t from the io ports
-  uint8_t in8(uint16_t port);
+  uint8_t in8(size_t port);
 
   // Send a uint8_t through the io ports
-  void out8(uint16_t port, uint8_t value);
+  void out8(size_t port, uint8_t value);
 
   // Get a uint16_t from the io ports
-  uint8_t in16(uint16_t port);
+  uint8_t in16(size_t port);
 
   // Send a uint16_t through the io ports
-  void out16(uint16_t port, uint16_t value);
+  void out16(size_t port, uint16_t value);
 
   // Get a uint32_t from the io ports
-  uint8_t in32(uint16_t port);
+  uint8_t in32(size_t port);
 
   // Send a uint32_t through the io ports
-  void out32(uint16_t port, uint32_t value);
+  void out32(size_t port, uint32_t value);
 };
