+++ src/rpi3/rpi3_io.cpp
@@ -20,37 +20,37 @@ void RPI3_IO_quark::reset(void)
 }
 
 // Get a uint8_t from the io ports
-uint8_t RPI3_IO_quark::in8(uint16_t port)
+uint8_t RPI3_IO_quark::in8(size_t port)
 {
   return *(volatile uint8_t *)(MMIO_BASE + port);
 }
 
 // Send a uint8_t through the io ports
-void RPI3_IO_quark::out8(uint16_t port, uint8_t value)
+void RPI3_IO_quark::out8(size_t port, uint8_t value)
 {
   *(volatile uint8_t *)(MMIO_BASE + port) = value;
 }
 
 // Get a uint16_t from the io ports
-uint16_t RPI3_IO_quark::in16(uint16_t port)
+uint16_t RPI3_IO_quark::in16(size_t port)
 {
   return *(volatile uint16_t *)(MMIO_BASE + port);
 }
 
 // Send a uint16_t through the io ports
-void RPI3_IO_quark::out16(uint16_t port, uint16_t value)
+void RPI3_IO_quark::out16(size_t port, uint16_t value)
 {
   *(volatile uint16_t *)(MMIO_BASE + port) = value;
 }
 
 // Get a uint32_t from the io ports
-uint32_t RPI3_IO_quark::in32(uint16_t port)
+uint32_t RPI3_IO_quark::in32(size_t port)
 {
   return *(volatile uint32_t *)(MMIO_BASE + port);
 }
 
 // Send a uint32_t through the io ports
-void RPI3_IO_quark::out32(uint16_t port, uint32_t value)
+void RPI3_IO_quark::out32(size_t port, uint32_t value)
 {
   *(volatile uint32_t *)(MMIO_BASE + port) = value;
 }
